Experiment with using C++ for detection.
Performance using C++ is a little better but not much. I tried doing a
multithreaded version in face_boxes_with_thread_pool, but didn't get
much speed boost there either.

My thought at this point is that provided that we can get the
tensorflow stuff working in C++, we may as well move over to the
language. I think it puts us in a much better place to start working
on performance.